requirements.txt
# Python 3.x
pyepics
anytree
